refactor completed with rl nets

+++ Solvers/NetSolvers/net_solver.py
@@ -17,7 +17,7 @@ class NetSolver(Solver):
             super().__init__(d)
             d_with_internals = np.zeros((self.m, self.m))
             d_with_internals[:self.n_taxa, :self.n_taxa] = self.d
-            self.d = torch.tensor(d_with_internals).to(self.device)
+            self.d = torch.tensor(d_with_internals).to(torch.float).to(self.device)
         self.net = net
         self.adj_mats = []
